Attempt to fix reaction caching
Usually discord.py waits until all guilds are available before
dispatching it's on_ready event, however if a guild is too slow it may
not be ready by this time. My suspicion is that this is why the
blurplefier info channel was not loaded in the prod bot. Let's see!